Override and append modes of file stream demonstrated.

+++ learning_tests/test/learning/java/io/_7_FileOutputStreamTest.java
@@ -9,39 +9,87 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 public class _7_FileOutputStreamTest {
+	private File file;
+	private FileOutputStream fileOutput;
+	private FileInputStream fileInput;
+	
+	@Before
+	public void setUp() {
+		file = new File("file");
+	}
+	
+	@After
+	public void tearDown() throws IOException {
+		if (fileOutput != null) {
+			fileOutput.close();
+		}
+		
+		if (fileInput != null) {
+			fileInput.close();
+		}
+		
+		if (file != null) {
+			file.delete();
+		}
+	}
 
 	@Test
 	public void bytesWrittenIntoFileOutputStreamAreReadFromFileInputStream() throws IOException {
-		File file = new File("file");
-		FileOutputStream fileOutput = new FileOutputStream(file);
+		fileOutput = new FileOutputStream(file);
 		
 		fileOutput.write(1);
 		fileOutput.write(2);
 		fileOutput.close();
 		
-		FileInputStream fileInput = new FileInputStream(file);
+		fileInput = new FileInputStream(file);
 		assertEquals(1, fileInput.read());
 		assertEquals(2, fileInput.read());
 		assertEquals("should be -1 after EOF", -1, fileInput.read());
 		assertEquals("should be -1 after EOF", -1, fileInput.read());
-		
-		fileInput.close();
-		file.delete();
 	}
 	
 	@Test
 	public void fileOutputStreamConstructorCreatesTheRealFileOnDisk() throws IOException {
-		File file = new File("file");
 		assertFalse("File should NOT be creted on disk", file.exists());
 		
-		FileOutputStream fileOutput = new FileOutputStream(file);
+		fileOutput = new FileOutputStream(file);
 		assertTrue("File should be creted on disk", file.exists());
-		
+	}
+	
+	@Test
+	public void writeAfterReopeningTheFileOverwritesByDefault() throws IOException {		
+		fileOutput = new FileOutputStream(file);
+		fileOutput.write(1);
 		fileOutput.close();
-		file.delete();
+		
+		// Reopen and rewrite
+		fileOutput = new FileOutputStream(file);
+		fileOutput.write(2);
+		
+		fileInput = new FileInputStream(file);
+		assertEquals(2, fileInput.read());
+		assertEquals("should be -1 after EOF", -1, fileInput.read());
+	}
+	
+	@Test
+	public void appendingNewBytesAfterReopeningTheFile() throws IOException {		
+		fileOutput = new FileOutputStream(file);
+		fileOutput.write(1);
+		fileOutput.close();
+		
+		// Reopen and rewrite
+		fileOutput = new FileOutputStream(file, true);
+		fileOutput.write(2);
+		
+		fileInput = new FileInputStream(file);
+		assertEquals(1, fileInput.read());
+		assertEquals("appended value", 2, fileInput.read());
+		assertEquals("should be -1 after EOF", -1, fileInput.read());
 	}
 
 }
